Refactor: Move State to application Use Stream instead of com.google.common.collect.Streams as it is not supported on Android.
