+++ application/src/main/java/bisq/application/State.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop;
+package bisq.application;
 
 public enum State {
     INITIALIZE_APP,
