+++ apps/desktop/desktop/src/main/java/bisq/desktop/DesktopController.java
@@ -17,6 +17,7 @@
 
 package bisq.desktop;
 
+import bisq.application.State;
 import bisq.bisq_easy.NavigationTarget;
 import bisq.chat.notifications.ChatNotificationService;
 import bisq.common.observable.Observable;
